Moved tests back to root.
Splited core foder and added to valhalla dir directly.
In valhalla.django.manage added new Python path
to be able to find the project.
