+++ .gitignore
@@ -2,4 +2,4 @@ venv/
 db.sqlite3
 *.py[cod]
 .idea/
-log/
+logs/
