+++ .travis.yml
@@ -14,7 +14,7 @@ before_install:
 
 install: pip install -r requirements.txt
 
-script: python core/manage.py test
+script: ./owtf.sh test
 
 branches:
   only:
