--- valhalla/core/__init__.py
@@ -1,2 +1,0 @@
-__author__ = 'owtf-valhalla'
-
