+++ valhalla/django/manage.py
@@ -2,9 +2,14 @@
 import os
 import sys
 
+# Add Valhalla to Python path.
+sys.path.append(os.path.dirname(os.path.realpath(__name__)))
+
+
 if __name__ == "__main__":
     os.environ.setdefault("DJANGO_SETTINGS_MODULE", "settings.settings")
 
     from django.core.management import execute_from_command_line
 
     execute_from_command_line(sys.argv)
+
